Fixed <0,0,0> issue

+++ Battleship/Battleship/rand_pool.h
@@ -11,16 +11,13 @@ public:
 	Point getNextRandPoint();
 	void removePointFromPool(Point pointToRemove);
 private:
-	//char ***myBoard;
 	std::vector < std::vector< std::vector<char> > > & myBoard;
 
 	// 1-dimension array containing potential moves
 	Point *potentialMoves;
 	int potentialMovesSize, rows, cols, depth;
 	std::vector < std::vector< std::vector<Point *> > > refTabel;
-	//Point ****refTabel;
 	std::vector < std::vector< std::vector<bool> > > isPointLegal;
-	//bool ***isPointLegal;
 	void initIsPointLegal();
 	void changeEnvalopPointsToFalse(std::vector < std::vector< std::vector<bool> > > &arr, Point p);
 	bool isPointPartOfShip(Point p);
@@ -28,8 +25,7 @@ private:
 
 	std::mt19937 engine{ 0 };
 #else
-	std::mt19937 engine{ time(NULL) };
+	std::mt19937 engine{ (unsigned int) time(NULL) };
 #endif
 	std::uniform_int_distribution<> dist;
 };
-
